Clean up

+++ src/main/java/nl/wiegman/homecontrol/services/service/StroomVerbruikService.java
@@ -59,7 +59,7 @@ public class StroomVerbruikService {
             } else {
                 Integer verbruik = meterstandRepository.getVerbruikInPeriod(periodeVan, periodeTotEnMet);
                 if (verbruik != null) {
-                    totaalVerbruikInKwh = verbruik.intValue();
+                    totaalVerbruikInKwh = verbruik;
                 }
             }
         }
